pyproject.toml
[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[project]
name = "clue-oda"
version = "1.0.0"
description = "Clue Optimized Data Aggregation."
dependencies = [
    "flask",
    "psycopg2",
    "click",
    "SQLAlchemy",
    "PyYAML",
    "pandas",
    "waitress",
]

[project.scripts]
clue = "clue_oda.data.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["clue_oda"]
